Moved plugins to their own files in src/plugins
